Move out `PUT` and `DELETE` methods for polls and add `GET` method for singular poll.
