+++ src/pages/api/polls/[id]/index.ts
@@ -1,0 +1,69 @@
+import withAuth, { AuthApiHandler } from '@/lib/auth';
+import prisma from '@/lib/prisma';
+
+const pollHandler: AuthApiHandler = async (req, res) => {
+  const pollId = req.query.id?.toString();
+  const poll = await prisma.poll.findFirst({
+    select: { id: true, schema: true, author: true },
+    where: { id: { equals: pollId } },
+  });
+
+  if (!poll) {
+    res.status(404).json({ message: '404 Not Found' });
+    return;
+  }
+
+  switch (req.method) {
+    case 'GET': {
+      res
+        .status(200)
+        .json({ poll: { id: poll.id, schema: JSON.parse(poll.schema) } });
+      break;
+    }
+
+    case 'PUT': {
+      if (poll.author.email !== req.session.user?.email) {
+        res.status(401).json({ message: '401 Forbidden' });
+        break;
+      }
+
+      const pollSchema = JSON.stringify(req.body.poll.schema);
+      if (!pollSchema) {
+        res.status(422).json({ message: '422 Unprocessable Entity' });
+        break;
+      }
+
+      const { count } = await prisma.poll.updateMany({
+        data: { schema: pollSchema },
+        where: { id: pollId },
+      });
+
+      res.status(200).json({ updated: count });
+      break;
+    }
+
+    case 'DELETE': {
+      if (poll.author.email !== req.session.user?.email) {
+        res.status(401).json({ message: '401 Forbidden' });
+        break;
+      }
+
+      const { count } = await prisma.poll.deleteMany({
+        where: {
+          id: req.body.poll.id,
+          author: { email: { equals: req.session.user?.email } },
+        },
+      });
+
+      res.status(200).json({ deleted: count });
+      break;
+    }
+
+    default: {
+      res.status(405).json({ message: '405 Method Not Allowed' });
+      break;
+    }
+  }
+};
+
+export default withAuth(pollHandler);
